Add spelling correction to text search queries
If a search has no results, then we will get suggestions from
Elasticsearch, re-query with the top suggestion, and return the results
from that query along with the suggestion that was used.

This change also changes search so that a document must match all terms
to be included in the results instead of any term.