Thread parent thread to child comments before serialization.
Normally, as a thread is being rendered, it goes through every child
comment (or whatever comments are in the pagination range) and
serializes them as it's properly ordering them, etc.

When these comments are serialized, they aren't fully materialized -
many relations are still unpopulated until the given fields are first
accessed.  For comments, a field that is accessed during serialization
in the 'comment_thread', or the ID of the comment's parent.

This operation, even if it uses the query cache, is still slow.  It goes
through reconstituing the object, which hits many mixins and other
helper code used to provide the ORM-like facilities over our usage of
MongoDB.

Since we always know the comment thread for a given comment, as in we
literally have the object available to us during serialization of the
thread, we now thread it through the various calls that are used to
reorder the child comments before everything is rendered to JSON.

This way, none of the magic that happens to lazy load this relation
occurs because the field is pre-populated, in essence.  This saves a lot
of work when rendering many comments in a given time slice of a thread.