Wait 10s after starting ElasticSearch.
Now that MongoDB 3.0 seems to download/untar faster than 2.6, we seem to
be right over the line of how quickly ES can start up and be listening
for connections, which leads the tests to try and connect too quickly,
which fails.

By sleeping for 10 seconds after starting ElasticSearch, it's almost a
certainty it will be up and running by the time we try to first connect
to it.