Verify that ES is configured properly before starting
Run the rake search:validate task before starting the service. If it
fails, then that means the mappings or indexes aren't in sync, the
former of which can cause corrupt data, and the latter, non-functioning
search. It fails with this exit code so that supervisor knows whether to
sleep before trying to restar the app or not.