feat: Adds additional data and filters to discussions API
This change adds new filters to the threads API. They are:
* Filtering by the thread type (discussion or question)
* Filtering by the thread author

In addition it also adds a new ``abuse_flagged_count`` field for threads. It
returns a count of the number of comments in a thread that are flagged for abuse.
This data is returned only if explicitly requested, so the platform can only
request it when needed for users with the requisite permissions.

Co-authored-by: Kshitij Sobti <kshitij@opencraft.com>