AA-303: Adding endpoint to return if a user has ever posted
This will check the Content table which stores both threads and
comments on threads. So if a user has created a post (thread) or commented
on one (comment), we will catch it.

The current use case for this is for the first Discussion milestone for
learners where we shhow them a celebration modal after their first
discussion post. In order to know if they have posted before, we hit
this endpoint.