feat: Add support for specifying tracking edit and close reasons
Adds support for specifiying an optional edit reason when editing a post which is preserved when the post is edited by a user that isn't the original author (for instance a moderator). The edit reason can be used by the platform to indicate why a post or comment was changed, for instance if it contained answers. It also stores the user that made this change and retains the previous version of the body in this case.

It also adds support for a post close reason code. This can be specified when closing the post. The user closing a post is also tracked as closed_by.