fix: add index for course id in user stats
The query to look up users by which ones have a course stat for a particular course is very slow when there are a large number of courses and users. This change adds an index for that field so that looking up users based on their course stats should be much faster.