fix: pagination for user stats api
The user stats API had a few issues that were breaking pagination. First of all the total count of entries was reported
as the count after limiting to the per-page limit. This meant that the count would be limited to the page count even
if there were more entries. This commit adds a facet to the query that fetches the total count, and the data after
limiting. Another issue was using skip after limit, which is the incorrect order. This commit fixes that.